more restriction

+++ main/views.py
@@ -1,21 +1,44 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from .models import Room
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 
 def home(request):
-    rooms = Room.objects.all()
-    context = {"rooms": rooms}
+    q = request.GET.get("q")
+    q = q if q else ""
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q)
+        | Q(name__icontains=q)
+        | Q(description__icontains=q)
+        | Q(host__username=q)
+    )
+
+    topics = Topic.objects.all()
+    context = {"rooms": rooms, "topics": topics, "rooms_count": rooms.count()}
     return render(request, "main/home.html", context)
 
 
 def room(request, pk):
     room = Room.objects.get(pk=pk)
-    context = {"room": room}
+    chats = room.message_set.all().order_by("-created")
+
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user, room=room, body=request.POST.get("body")
+        )
+        return redirect("room", pk=room.id)
+    context = {"room": room, "chats": chats}
     return render(request, "main/room.html", context)
 
 
+@login_required(login_url="/login")
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -27,9 +50,13 @@ def create_room(request):
     return render(request, "main/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You don't have the permission to do this! ")
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -39,10 +66,60 @@ def update_room(request, pk):
     return render(request, "main/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
     context = {"room": room}
+
+    if request.user != room.host:
+        return HttpResponse("You don't have the permission to do this! ")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
     return render(request, "main/delete.html", context)
+
+
+def login_view(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exists!")
+        user = authenticate(request, username=username, password=password)
+        if user:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username OR password is incorrect!")
+    context = {"page": page}
+    return render(request, "main/login_register.html", context)
+
+
+def logout_view(request):
+    logout(request)
+    return redirect("home")
+
+
+def register_view(request):
+    page = "register"
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occured during registration")
+    context = {"page": page, "form": form}
+    return render(request, "main/login_register.html", context)
